+++ packages/to-json-schema/src/functions/toJsonSchema/toJsonSchema.ts
@@ -6,6 +6,7 @@ import type {
   ConversionContext,
   JsonSchema,
 } from '../../types/index.ts';
+import { ReferenceMap } from '../../utils/index.ts';
 
 /**
  * Converts a Valibot schema to the JSON Schema format.
@@ -22,7 +23,7 @@ export function toJsonSchema(
   // Initialize JSON Schema context
   const context: ConversionContext = {
     definitions: {},
-    referenceMap: new Map(),
+    referenceMap: new ReferenceMap(),
     getterMap: new Map(),
   };
 
@@ -33,8 +34,6 @@ export function toJsonSchema(
   if (definitions) {
     for (const key in definitions) {
       context.referenceMap.set(definitions[key], key);
-      // Reserve every key because the same schema can have multiple names.
-      context.definitions[key] = {};
     }
     for (const key in definitions) {
       context.definitions[key] = convertSchema(